LGTM

+++ src/services/accounts.service.ts
@@ -1,6 +1,7 @@
 import { Prisma, Account } from '@prisma/client'
 import createError from 'http-errors'
 import CreateAccountDto from '../dtos/accounts/req/create-account.dto'
+import UpdateAccountDto from '../dtos/accounts/req/update-account.dto'
 import prisma from './prisma.service'
 
 export default class AccountsService {
@@ -12,8 +13,11 @@ export default class AccountsService {
     return prisma.account.findUnique({ where: { id } })
   }
 
-  static async update(id: string, input: string): Promise<Account> {
+  static async update(id: string, input: UpdateAccountDto): Promise<Account> {
     try {
+      if (!id) {
+        throw new createError.UnprocessableEntity('bad request')
+      }
       return prisma.account.update({
         data: input,
         where: {
@@ -46,6 +50,9 @@ export default class AccountsService {
   }
 
   static async delete(id: string): Promise<Account> {
+    if (!id) {
+      throw new createError.UnprocessableEntity('bad request')
+    }
     const account = await prisma.account.delete({
       where: {
         id,
